ww_finance-ai/requirements.txt
# 核心依赖
langchain==0.1.20
langgraph==0.2.0
openai==1.30.0

# 数据处理
pandas==2.2.0
numpy==1.26.3

# API服务
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.0
pydantic-settings==2.1.0

# 工具
python-dotenv==1.0.0
requests==2.31.0
